+++ .gitignore
@@ -7,7 +7,7 @@ node_modules
 
 # testing
 /coverage
-
+.vscode
 # production
 /build
