+++ backend/index.js
@@ -1,10 +1,11 @@
 import express from 'express';
 import mongoDB from './db.js';
-import userRoutes from './Routes/CreatUser.js';
+import userRoutes from './Routes/CreateUser.js';
 import displayData from './Routes/DisplayData.js';
 import orderData from './Routes/OrderData.js';
 import cors from 'cors';
-
+import adminRoute from './Routes/admin/CreateAdmin.js';
+import homeRoutes from './Routes/admin/HomeAPi.js';
 const app = express();
 const port = 5000;
 
@@ -15,6 +16,10 @@ mongoDB();
 app.use(cors({ origin: 'http://localhost:3000' }));
 
 app.use(express.json());
+// Admin Apis
+app.use("/api/admin",adminRoute);
+app.use('/api/admin/home', homeRoutes);
+// Normal Apis
 app.use('/api', userRoutes);
 app.use('/api', displayData);
 app.use('/api', orderData);
